Added the ability to set content shareability and visibility. Added some user access control.

+++ app/controllers/ContentController.scala
@@ -5,14 +5,13 @@ import service.{ContentManagement, Authentication}
 import models.Content
 
 /**
- * Created with IntelliJ IDEA.
- * User: camman3d
- * Date: 3/22/13
- * Time: 1:27 PM
- * To change this template use File | Settings | File Templates.
+ * The controller for dealing with content.
  */
 object ContentController extends Controller {
 
+  /**
+   * Action mix-in to get the content from the request
+   */
   def getContent(id: Long)(f: Content => Result)(implicit request: Request[_]) = {
     val content = Content.findById(id)
     if (content.isDefined) {
@@ -21,12 +20,18 @@ object ContentController extends Controller {
       NotFound
   }
 
+  /**
+   * Content creation page
+   */
   def createPage = Authentication.authenticatedAction() {
     implicit request =>
       implicit user =>
         Ok(views.html.content.create())
   }
 
+  /**
+   * Creates content based on the posted data
+   */
   def create = Authentication.authenticatedAction(parse.multipartFormData) {
     implicit request =>
       implicit user =>
@@ -45,6 +50,9 @@ object ContentController extends Controller {
         Redirect(routes.Application.home()).flashing("success" -> "Content added")
   }
 
+  /**
+   * Content view page
+   */
   def view(id: Long) = Authentication.authenticatedAction() {
     implicit request =>
       implicit user =>
@@ -54,21 +62,69 @@ object ContentController extends Controller {
         }
   }
 
+  /**
+   * Content deletion endpoint
+   */
   def delete(id: Long) = Authentication.authenticatedAction() {
     implicit request =>
       implicit user =>
         getContent(id) {
           content =>
-            content.delete()
-            Redirect(routes.ContentController.mine()).flashing("success" -> "Content deleted.")
+
+            // Make sure the user is able to edit
+            if (content isEditableBy user) {
+              content.delete()
+              Redirect(routes.ContentController.mine()).flashing("success" -> "Content deleted.")
+            } else
+              Forbidden
         }
   }
 
+  /**
+   * "My Content" page
+   */
   def mine = Authentication.authenticatedAction() {
     implicit request =>
       implicit user =>
         Ok(views.html.content.mine())
   }
 
+  /**
+   * Set content visibility endpoint
+   */
+  def setVisibility(id: Long) = Authentication.authenticatedAction(parse.urlFormEncoded) {
+    implicit request =>
+      implicit user =>
+        getContent(id) {
+          content =>
+
+            // Make sure the user is able to edit
+            if (content isEditableBy user) {
+              val visibility = request.body("visibility")(0).toInt
+              content.copy(visibility = visibility).save
+              Redirect(routes.ContentController.view(id)).flashing("success" -> "Visibility updated.")
+            } else
+              Forbidden
+        }
+  }
+
+  /**
+   * Set content shareability endpoint
+   */
+  def setShareability(id: Long) = Authentication.authenticatedAction(parse.urlFormEncoded) {
+    implicit request =>
+      implicit user =>
+        getContent(id) {
+          content =>
+
+            // Make sure the user is able to edit
+            if (content isEditableBy user) {
+              val shareability = request.body("shareability")(0).toInt
+              content.copy(shareability = shareability).save
+              Redirect(routes.ContentController.view(id)).flashing("success" -> "Shareability updated.")
+            } else
+              Forbidden
+        }
+  }
 
 }
